# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HEIC to JPG Conversion API
  version: 1.0.0
  description: |
    Free HEIC/HEIF to JPEG conversion API.
    
    - Send a multipart/form-data POST with a `file` field containing a HEIC/HEIF image.
    - The API returns an `image/jpeg` binary response.
    - CORS is enabled for all origins.
servers:
  - url: https://heic-to-jpg.io
    description: Production
paths:
  /api/convert:
    options:
      summary: CORS preflight
      responses:
        '204':
          description: No Content
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
              description: CORS header
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
    post:
      summary: Convert HEIC/HEIF to JPEG
      description: |
        Accepts a multipart/form-data payload with a single `file` field.
        Returns a JPEG image. Output filename mirrors the input with `.jpg` extension.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: HEIC/HEIF image
      responses:
        '200':
          description: JPEG image
          headers:
            Content-Type:
              schema:
                type: string
              description: image/jpeg
            Content-Disposition:
              schema:
                type: string
              description: Suggested filename for the JPEG
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '400':
          description: Bad Request (missing file or content type)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '422':
          description: Unprocessable Entity (decode failure)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
components: {} 